pyproject.toml
[project]
name = "network-king"
version = "0.1.0"
description = "A mobile-first networking game web app built with FastAPI."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "jinja2>=3.1.0,<4.0.0",
  "python-multipart>=0.0.9,<1.0.0",
  "qrcode[pil]>=7.4.2,<8.0.0",
  "reportlab>=4.2.0,<5.0.0",
  "sqlalchemy>=2.0.30,<3.0.0",
  "uvicorn[standard]>=0.30.0,<1.0.0",
]

[project.optional-dependencies]
gcp = [
  "google-cloud-storage>=2.16.0,<3.0.0",
]
test = [
  "httpx>=0.27.0,<1.0.0",
  "pytest>=8.2.0,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]

